requirement.txt
Django==2.2.12
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
uWSGI==2.0.15
